MB-67520: Use uint64_t for distributionKey
The "distribution key" is used to allocate a BgFetcher to a vBucket,
when there are more BgFetchers (Reader threads) than vBuckets (active or
replica).

  int workersForVBucket = ...;
  int workerIndex = hash(distributionKey) % workersForVBucket;

The CAS is used as distributionKey. The old code used 32-bit int, so it
lost some of the bits, resulting in potentially less uniform allocation.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230509
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/ep_bucket.cc
@@ -2810,7 +2810,7 @@ std::ostream& operator<<(std::ostream& os, const EPBucket::FlushResult& res) {
     return os;
 }
 
-BgFetcher& EPBucket::getBgFetcher(Vbid vbid, uint32_t distributionKey) {
+BgFetcher& EPBucket::getBgFetcher(Vbid vbid, uint64_t distributionKey) {
     const auto numBgFetchers = bgFetchers.size();
     const auto numActiveVBuckets = vbMap.getVBStateCount(vbucket_state_active);
     const auto bgFetcher = selectWorkerForVBucket(
